# openapi.yaml
openapi: "3.1.1"
info:
  title: Conduit CE by AppSynergy — API
  version: "1.0.0"
  summary: Community Edition — remote infrastructure management platform API
  description: |
    Conduit is a remote infrastructure management platform that provides secure,
    instant shell access to any machine — without inbound ports, SSH keys, VPNs,
    or trusting the network.

    ## Security Framework

    This API implements security controls aligned with:
    - **NIST SP 800-53 Rev. 5** — Security and Privacy Controls
    - **NIST SP 800-63B** — Digital Identity Guidelines (AAL2 baseline, AAL3 with hardware authenticator)
    - **NIST SP 800-131A Rev. 2** — Cryptographic Transitions
    - **NIST SP 800-57** — Key Management
    - **FIPS 203 (ML-KEM)** — Post-Quantum Key Encapsulation (X25519MLKEM768 hybrid TLS)
    - **FIPS 204 (ML-DSA)** — Post-Quantum Digital Signatures (JWT signing, future)
    - **FIPS 205 (SLH-DSA)** — Post-Quantum Stateless Hash-Based Signatures (binary signing)

    ## Transport Security

    - **TLS 1.3 mandatory** on all connections. TLS 1.0, 1.1, 1.2 are disabled.
    - **X25519MLKEM768 hybrid PQC** key exchange on all TLS connections.
    - **QUIC primary** (UDP 443) with automatic WebSocket/HTTP fallback (TCP 443).
    - Cipher suites: TLS_AES_256_GCM_SHA384, TLS_AES_128_GCM_SHA256, TLS_CHACHA20_POLY1305_SHA256.

    ## Forbidden Algorithms (NIST SP 800-131A Disallowed)

    MD5, SHA-1, DES, 3DES, RC4, TLS 1.0/1.1/1.2, non-CSPRNG sources.

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /agents/{agentId}/shell/sessions:
    get:
      operationId: listShellSessions
      summary: List shell sessions on agent
      description: |
        Lists shell sessions for the specified agent. Includes active, detached,
        and optionally closed sessions. Detached sessions are resumable from any
        browser or platform.

        **NIST AU-2:** Session listing provides visibility into active access.
        **NIST AC-3:** Only sessions owned by the requesting user or visible to
        users with org_admin+ role are returned.
      tags: [Shell]
      x-nist-controls: [AU-2, AC-3]
      x-owasp: [A01, API1, API3]
      parameters:
        - $ref: "#/components/parameters/PathAgentId"
        - name: status
          in: query
          description: Filter by session status. Omit to return active + detached.
          schema:
            type: string
            enum: [active, detached, closed, all]
        - name: pinned
          in: query
          description: Filter by pinned status
          schema:
            type: boolean
      responses:
        "200":
          description: Shell sessions
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ShellSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createShellSession
      summary: Create a new shell session
      description: |
        Opens a new PTY shell session on the agent.
        Multiple concurrent sessions per agent are supported via CWP multiplexing.
        Session data flows over the WebSocket at `/agents/{agentId}/shell/sessions/{sessionId}/ws`.

        Sessions are persistent by default — if the browser disconnects, the PTY
        remains alive on the agent and the session enters `detached` status.
        Reconnect from any platform via the attach WebSocket endpoint.

        Set `pinned: true` for sessions monitoring long-running background tasks
        that must remain accessible indefinitely regardless of idle timeout.

        **NIST AU-2, AU-3:** Shell session start logged with user, agent, timestamp.
        **NIST AC-12:** Idle timeout enforced on detached sessions (unless pinned).
        **x-audit-event:** shell.session.started
      tags: [Shell]
      x-nist-controls: [AU-2, AU-3, AC-12]
      x-owasp: [A01, A03, API1, API5]
      x-audit-event: shell.session.started
      parameters:
        - $ref: "#/components/parameters/PathAgentId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cols:
                  type: integer
                  default: 80
                  minimum: 1
                  maximum: 10000
                  description: Terminal width in columns
                rows:
                  type: integer
                  default: 24
                  minimum: 1
                  maximum: 10000
                  description: Terminal height in rows
                shell:
                  type: string
                  maxLength: 4096
                  description: Shell to launch (e.g., /bin/bash, cmd.exe). Defaults to OS default.
                recording:
                  type: boolean
                  default: true
                  description: Enable terminal session recording (asciicast v2)
                pinned:
                  type: boolean
                  default: false
                  description: |
                    Pin session to prevent idle timeout. Pinned sessions stay alive
                    indefinitely when detached. Use for monitoring long-running tasks.
                idleTimeout:
                  type: integer
                  minimum: 60
                  maximum: 86400
                  default: 3600
                  description: |
                    Seconds of detached inactivity before automatic closure.
                    Ignored when pinned is true. Minimum 60 seconds.
              additionalProperties: false
      responses:
        "201":
          description: Shell session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShellSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          description: Concurrent session limit reached
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token signed with Ed25519 (FIPS 204 ML-DSA-65 when Go stdlib ships it).
        Obtained via WebAuthn login, CLI device flow, or CI token.

        **JWT Claims:**
        - `sub` — User ID (UUID v4)
        - `tid` — Tenant ID (determines data scope)
        - `services` — Service slugs the tenant has enabled (e.g. `["remote-access"]`)
        - `roles` — Role identifiers within the tenant
        - `permissions` — Permission strings
        - `iat` / `exp` — Issued/expiration timestamps

        Access tokens expire after 15 minutes (configurable).
        Use refresh token to obtain new access tokens.

        **NIST IA-5, SC-13, SC-23.**
      x-nist-controls: [IA-5, SC-13, SC-23]

    AgentToken:
      type: http
      scheme: bearer
      bearerFormat: HMAC-SHA256
      description: |
        Agent authentication token. HMAC-SHA256, per-agent.
        Presented during WebSocket upgrade (`Authorization: Bearer <token>`)
        or in the first CWP HELLO frame over QUIC.

        **NIST IA-2, IA-5.**

    CiToken:
      type: apiKey
      in: header
      name: Authorization
      description: |
        CI/automation token. Scoped, revocable.
        Set as `CONDUIT_TOKEN` environment variable.
...
  parameters:
...
    PathAgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: "401 — Authentication required (NIST IA-2)"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
      headers:
        X-Request-Id:
          $ref: "#/components/headers/X-Request-Id"
        X-Content-Type-Options:
          $ref: "#/components/headers/X-Content-Type-Options"
        Strict-Transport-Security:
          $ref: "#/components/headers/Strict-Transport-Security"
    Forbidden:
      description: "403 — Insufficient permissions (NIST AC-3, AC-6)"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
      headers:
        X-Request-Id:
          $ref: "#/components/headers/X-Request-Id"
        X-Content-Type-Options:
          $ref: "#/components/headers/X-Content-Type-Options"
        Strict-Transport-Security:
          $ref: "#/components/headers/Strict-Transport-Security"
    NotFound:
      description: "404 — Resource not found"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
      headers:
        X-Request-Id:
          $ref: "#/components/headers/X-Request-Id"
        X-Content-Type-Options:
          $ref: "#/components/headers/X-Content-Type-Options"
        Strict-Transport-Security:
          $ref: "#/components/headers/Strict-Transport-Security"
...
  headers:
    X-Request-Id:
      description: "Request correlation ID for audit trail (NIST AU-3)"
      schema:
        type: string
        format: uuid
    X-Content-Type-Options:
      description: "Prevents MIME type sniffing (NIST SC-18)"
      schema:
        type: string
        const: nosniff
    Strict-Transport-Security:
      description: "Enforce HTTPS (NIST SC-8). max-age=63072000 (2 years)."
      schema:
        type: string
        example: "max-age=63072000; includeSubDomains; preload"

  # ──────────────────────────────────────────────────────────────
  # SCHEMAS
  # ──────────────────────────────────────────────────────────────
  schemas:
...
    ProblemDetail:
      type: object
      description: "RFC 9457 Problem Details for HTTP APIs"
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        traceId:
          type: string
          description: "Distributed tracing ID (NIST AU-3)"
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string

    # ── Pagination ──
...
    ShellSession:
      type: object
      required: [id, agentId, userId, status, pinned, createdAt]
      properties:
        id:
          type: string
          format: uuid
        agentId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        agentHostname:
          type: string
          description: Hostname of the agent (denormalized for display)
        userEmail:
          type: string
          format: email
          description: Email of the session owner (denormalized for display)
        status:
          type: string
          enum: [active, detached, closed]
          description: |
            - `active` — browser attached, I/O flowing
            - `detached` — no browser attached, PTY alive on agent, output buffered
            - `closed` — PTY exited or session terminated
        shell:
          type: string
        cols:
          type: integer
        rows:
          type: integer
        recording:
          type: boolean
        pinned:
          type: boolean
          default: false
          description: |
            Pinned sessions never idle-timeout. Use for monitoring long-running
            background tasks that must remain accessible from any platform.
            NIST AC-12: pinned sessions bypass idle reaper but are still subject
            to explicit termination and JWT re-authentication on attach.
        idleTimeout:
          type: integer
          minimum: 0
          maximum: 86400
          default: 3600
          description: |
            Seconds of detached inactivity before the session is automatically
            closed. 0 means use server default. Ignored when pinned is true.
            NIST AC-12: enforces session termination after inactivity.
        websocketUrl:
          type: string
          description: Relative WebSocket URL for attaching to this session
        createdAt:
          type: string
          format: date-time
        detachedAt:
          type:
            - string
            - "null"
          format: date-time
          description: Timestamp when session was last detached (browser disconnected)
        closedAt:
          type:
            - string
            - "null"
          format: date-time
...